ar71xx: add dts files for certain devices
The flash partitions mentioned in the dts are here for having
an OF node handle to reference them for MAC-address-loading and
EEPROM-loading.
Partitions are actually detected in runtime using "linux,part-probe = "tp-link""
The offsets are not read, it's just a OF node <-> partition label mapping

Also, for eth0 at the WR74x devices: the MAC-address is decremented by one!
> mtd-mac-address-increment = <(-1)>;
The code actually reads this as an "u32", doing the math on one field of "u8 mac[6]"
will result in overflow, and finally -1.
Although the result is ok, is this approach the correct way?